requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
certifi==2022.9.24
channels==4.0.0
channels-redis==4.0.0
cloudinary==1.30.0
colorama==0.4.4
dj-database-url==1.0.0
Django==4.0.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gunicorn==20.1.0
msgpack==1.0.4
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.5
pyparsing==3.0.9
pytz==2022.6
redis==4.3.5
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.13
whitenoise==6.2.0
